feat(Market): creditPrices are now not doubles but integers now.

+++ src/main/java/de/uniks/stp24/model/SeasonComponent.java
@@ -4,6 +4,6 @@ public record SeasonComponent(
         String transActionTypeText,
         String resourceType,
         int resourceAmount,
-        double moneyAmount
+        int moneyAmount
 ) {
 }
